Updated to use BigDecimal isntead of double to avoid rounding errors

+++ estore-api/src/main/java/com/estore/api/estoreapi/persistence/UserDAO.java
@@ -1,6 +1,7 @@
 package com.estore.api.estoreapi.persistence;
 
 import java.io.IOException;
+import java.math.BigDecimal;
 
 import com.estore.api.estoreapi.users.User;
 import com.estore.api.estoreapi.users.Admin;
@@ -97,7 +98,7 @@ public interface UserDAO {
      * 
      * @throws IOException if an issue with underlying storage
      */
-    Boolean addToCart(String username, int beefId, double weight) throws IOException;
+    Boolean addToCart(String username, int beefId, BigDecimal weight) throws IOException;
 
 
     /**
